teraarc api/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
python-dotenv==1.0.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
gunicorn==21.2.0
python-jose==3.3.0
